Split out dispatch tests to its own class
